backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
flask-mail==0.10.0
flask-cors==4.0.1
flask-socketio==5.3.6
flask-bcrypt==1.0.1
flask-limiter==3.8.0
flask-restx==1.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
reportlab==4.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.8
twilio==9.3.2
Pillow==10.4.0
marshmallow==3.21.3
email-validator==2.2.0
cryptography==43.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
